Combine lock tests into locks_test.py
